mejorando validaciones

+++ PracticaIntegradora/src/daos/dbManager/cart.dao.js
@@ -7,33 +7,77 @@ class CartDao {
   }
 
   async findCartById(_id) {
-    return await cartModel.findById(_id);
+    try {
+      if (mongoose.Types.ObjectId.isValid(_id)) {
+        const cartFound = await cartModel.findById({ _id });
+
+        if (cartFound) {
+          return await cartModel.findById(_id);
+        }
+        return "Cart not found";
+      }
+      console.log("Formato de id no válido");
+    } catch (error) {
+      console.log(error);
+    }
   }
 
   async createCart(cart) {
-    return await cartModel.create(cart);
+    try {
+      if (productModel.isValid) {
+        return await cartModel.create(cart);
+      }
+
+      console.log("Hay campos incompletos");
+      return false;
+    } catch (error) {
+      console.log(error);
+    }
   }
 
   async updateCart(_id, cart) {
-    return await cartModel.findByIdAndUpdate({ _id }, cart);
+    try {
+      if (mongoose.Types.ObjectId.isValid(_id)) {
+        const cartFound = await cartModel.findById({ _id });
+
+        if (cartFound) {
+          return await cartModel.findByIdAndUpdate({ _id }, cart);
+        }
+        return "Cart not found";
+      }
+      console.log("Formato de id no válido");
+    } catch (error) {
+      console.log(error);
+    }
   }
 
   async deleteCart(_id) {
+    //revisar si funciona
 
-    //hay que ver sis funciona este
-    await productModel.deleteMany({ product: _id });
+    try {
+      if (mongoose.Types.ObjectId.isValid(_id)) {
+        const cartFound = await cartModel.findById({ _id });
 
-    return await cartModel.findByIdAndDelete({ _id });
+        if (cartFound) {
+          await productModel.deleteMany({ product: _id });
+
+          return await cartModel.findByIdAndDelete({ _id });
+        }
+        return "Cart not found";
+      }
+      console.log("Formato de id no válido");
+    } catch (error) {
+      console.log(error);
+    }
   }
 
-   errorMessage(error){
-     console.log(error);
-     res.json({
-       error,
-       message: "Error",
-     });
+  errorMessage(error) {
+    console.log(error);
+    res.json({
+      error,
+      message: "Error",
+    });
   }
 }
-
 
 export default new CartDao();
